requirements.txt
black==23.11.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
filelock==3.13.1
identify==2.5.33
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pre-commit==3.5.0
PyYAML==6.0.1
ruff==0.1.7
tomli==2.0.1
typing_extensions==4.9.0
virtualenv==20.25.0
django==4.2.9
psycopg2-binary>=2.8
django-environ==0.11.2
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
beautifulsoup4==4.12.3
llama_index==0.9.33
uuid==1.30
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
gevent==23.9.1
gevent==23.9.1
pybloom-live==4.0.0
django-advanced-filters==2.0.0
